refactor: convert useTheme to React Context for synchronized state
Fixes theme synchronization issue where chart components wouldn't
update their Chart.js colors when toggling dark mode. The previous
implementation used local useState in each component, creating
independent state instances.

Changes:
- Create ThemeContext and ThemeProvider in src/context/ThemeContext.tsx
- Move theme logic from hook to context with shared state
- Wrap App with ThemeProvider for global theme state
- Update all components to import useTheme from @/context
- Remove old useTheme.ts hook file

Now when ThemeToggle changes the theme, all components (including
ProgressChart and WeightChart) will immediately update with the new
theme colors.
