+++ src/components/ui/ThemeToggle.tsx
@@ -1,4 +1,4 @@
-import { useTheme } from '@/hooks/useTheme';
+import { useTheme } from '@/context';
 
 export function ThemeToggle() {
   const { theme, toggleTheme } = useTheme();
