+++ src/context/index.ts
@@ -1,1 +1,2 @@
 export { AppProvider, useApp } from './AppContext';
+export { ThemeProvider, useTheme } from './ThemeContext';
