+++ src/hooks/index.ts
@@ -1,3 +1,2 @@
 export { useChart } from './useChart';
 export { useLocalStorage } from './useLocalStorage';
-export { useTheme } from './useTheme';
